Fix header ordering dependency
There are types used in `decode.h` and `perfTimer.h` that are defined in
`windowsmediaphoto.h` which is included (in most cases) transitively by
`strcodec.h`, leading to an unfortunate ordering dependency for header
includes.

Running `clang-format` resulted in headers being ordered
alphabetically, which broke on this dependency. This commit restores
enough of the original ordering to get things building again. (A better
fix would be to require `windowsmediaphoto.h` directly in the headers
that use the types defined therein, but this whole mess is so fragile
that I've erred on the side of changing the least to get things building
again.)