Add more C++ wrappers for decoding
Three classes that work in concert with each other are added here.
`ImageDecoder` wraps the relevant C structs and function calls to do
decoding with JXRLib. `FormatConverter` wraps the struct used by JXRLib
to decode from one format into another. Finally, the `CodecFactory`
class is used to create and initialize the other two, and is a `friend`
to both to handle cases where structs from one are needed to initialize
other structs.