Add a simple test of C++ wrappers
The `jxrdecode` app is a direct port of JXRLib's `JxrDecApp.c`, minus
argument handling. This exists simply as a means of quickly validating
the C++ wrappers are functioning as expected.The `jxrdecode` app is a
direct port of JXRLib's `JxrDecApp.c`, minus argument handling. This
exists simply as a means of quickly validating the C++ wrappers are
functioning as expected.