Move Encoder/Decoder release to close method
With the `Release` method in the destructor, when the `ImageDecoder`
class is destroyed it also attempts to release the `WMPStream` held by
the underlying `ImageDecode` C-struct. When an `ImageDecoder` C++ object
is copied, both copies will point to the same underlying C-struct, and
when the first is destroyed the stream is closed for both.

One potential solution would be to use C++ copy/move constructors in
combination with the rudimentary `WMPStream` ownership tracking
implemented in the `ImageDecode` C-struct, but this is complicated and
would require significant testing/validation.

This alternative, compromise solution simply moves the call to `Release`
into an instance method, requiring the consumer of a `ImageDecoder`
object to manually close the underlying stream.