Remove buggy SWIG Exception translation
I was operating under the (erroneous) assumption that SWIG would handle
wrapping of methods that might catch a C++ exception and translating
that exception to the Java equivalent if no custom method body was
provided. Instead, SWIG simply replaced those methods with an empty
body. For now, we'll just allow C++ exceptions to bubble up, but for
future we should figure out a better way to catch and re-throw.