Change from unsigned to signed char for bytes
Due to the absence of unsigned bytes in Java, if we use `unsigned char`
as the type for C++, SWIG will translate this as `short` in Java (as
this is the smallest type that can support the full range of values for
`unsigned char`s). However, as we don't care so much about the value of
the bytes as the bits themselves, it's annoying that we have to
translate back and forth to `short` just to get this all working.