Switch to using NIO byte buffers
Using an NIO byte buffer allows us to both not incur the copy overhead
of the JNI function `GetByteArrayElements()` and use the builtin SWIG
type maps in order that we might utilise `unsigned char *` throughout.

The semantics of SWIG and Java byte array handling also mean that
`ReleaseByteArrayElements()` is called at the end of the JNI call
releasing it to the JVM to do with as it might desire.  This fact is
likely the root cause of our rather stuborn decode failures and
segmentation faults performing in memory decodes when passing in a
Java byte array.

NOTE: Using the included SWIG interface file requires SWIG 3.0.3+