Overhaul memory management semantics
In introducing an `ImageDecoder` destructor it became clear that the
`Closeable` was no longer needed.  Also, returning an `ImageDecoder`
from `CodedFactory::decoderFromBytes()` was playing havoc with the Java
JNI bindings destroying their C++ class counterparts.  This was in fact
causing a double free to be attempted on the `ImageDecoder` itself.

`Decode` no longer implements `Closable` as it no longer has to and
there are now corresponding destructors in the C++ layer to clean up the
C layer.  It is also expected that an `ImageDecoder` be allocated and
passed in to `CodedFactory::decoderFromBytes()` rather than the method
allocating one for you and returning it.