Use `memmove()` rather than `memcpy()` to aid older GLIBC compatibility
Around the release of GLIBC 2.13 changes were made to the behaviour of
`memcpy()` and `memmove()` during some 64-bit Intel processor specific
improvements.  This broke certain applications that relied, knowingly or
not, on the previously "undefined" behaviour creating somewhat of an
Ulrich vs. Linus shitstorm:

 * https://sourceware.org/bugzilla/show_bug.cgi?id=12518

As such, there are two symbol versions that we can target with GLIBC
2.14+:

 * memcpy@GLIBC_2.2.5
 * memcpy@GLIBC_2.14

The symantics here are certainly complicated but end the result is the
same: we have an CentOS 6 deployment target, which ships with GLIBC
2.12, and thus can only use the GLIBC 2.2.5 symbol.  As Linus repeatedly
argued on the aforementioned bug, there's really very little reason not
to be using `memmove()`, it's both safer and the performance penalties
are negligable when not working with overlapping areas of memory.