Optimise for size via `-Os`
In implementing 90977d3 and looking at the dynamic symbol table of the
resulting shared object the memcpy@GLIBC_2.14 symbol is still referenced
when linking against GLIBC 2.14+ even though it is unused.  This is
because "string.h" exports that symbol and in most cases the compiler
defaults to `-O0` causing "dead" symbols to remain in the dynamic symbol
table.