Test offset and length version of the API
This commit provides an additional `--in-memory-offset` flag which
advanced the buffer 1MB before reading values from the source file.  As
printing all the bytes was getting a bit obnoxious a switch to printing
just an MD5 hash digest of the source and decoded bytes was performed.
This is in line with the Java tests so it is now easy to visually
compare the two implementations for differences at a byte level at a
glance.

NOTE: OpenSSL headers and `-lcrypto` are now required for the C++ binary