Enable decoding of images to Java byte arrays
This skips one-half of the `ByteBuffer` dance that we currently have to
do to get JXR data to C++ and the decoded image bytes back to Java.
Currently, this requires manual edits to the SWIG generated Java and C++
files, so this commit also checks these in. As soon as we can re-write
the SWIG interface file, we should be able to re-exclude these files.