Unify API for NIO Buffers and byte[]
This is the first step of refactoring the C++ and Java APIs so that they are
more consistent. On the C++ side, input buffers are always `unsigned char` and
instead of multiple overloads, optional offsets are represented as optional
parameters. This does necessitate rearranging parameter order on some methods,
so all methods were re-arranged so that offsets and lengths appear at the end of
the parameter list.

On the Java side, instead of relying on the built-in SWIG mechanisms for passing
NIO Buffers and `byte[]` across the JNI bridge, we have copied the definitions
into our `JXR.i` definitions and changed the `byte[]` implementation to pass
across as `unsigned char *`. This allows us to use the same C++ methods for
both, though due to limitations with SWIG auto-generation, we need to use the
`%rename` directive to prevent definitions from replacing each other. This
results in a new set of more descriptive method names.